[update]爆発にエフェクト追加

+++ SirensMoon/DamageCut.cpp
@@ -33,9 +33,12 @@ void DamageCut::Update() {
 
 void DamageCut::Render() {
 	if (_visible && _lifetime < 59) {
-		DrawBox(0, 0, screen_W, _pos.y, GetColor(120, 20, 30), 1);
-		DrawBox(0, _pos.y + _size.y, screen_W, screen_H, GetColor(120, 20, 30), 1);
-		DrawExtendGraph(_pos.x - _moveX, _pos.y, _pos.x + _size.x - _moveX, _pos.y + _size.y, _cg, 0);
+		DrawBox(0, 0, screen_W, static_cast<int>(_pos.y), GetColor(120, 20, 30), 1);
+		DrawBox(0, static_cast<int>(_pos.y + _size.y), screen_W, screen_H, GetColor(120, 20, 30), 1);
+		DrawExtendGraph(static_cast<int>(_pos.x - _moveX),
+			static_cast<int>(_pos.y),
+			static_cast<int>(_pos.x + _size.x - _moveX),
+			static_cast<int>(_pos.y + _size.y), _cg, 0);
 	}
 }
